Исправляен косяк с @EqualsAndHashCode.Exclude и @ToString.Exclude в Book.comments. Исправлен косяк с инициализацией коллекций в JpaBookRepository.java. Тестовые классы перенесены в пакеты основных классов - чтобы не было бардака.
