+++ hw06/src/test/java/ru/otus/hw/base/ConfigurableByPropertiesTestBase.java
@@ -1,4 +1,4 @@
-package base;
+package ru.otus.hw.base;
 
 import org.springframework.test.context.TestPropertySource;
