+++ hw06/src/test/java/ru/otus/hw/data/CommentsArgProvider.java
@@ -1,4 +1,4 @@
-package data;
+package ru.otus.hw.data;
 
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.provider.Arguments;
@@ -9,19 +9,18 @@ import java.util.List;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
-public class BooksArgProvider implements ArgumentsProvider {
+public class CommentsArgProvider implements ArgumentsProvider {
     private final List<Book> books;
-
-    public BooksArgProvider() {
+    public CommentsArgProvider() {
         var provider = new TestDataProvider();
         this.books = provider.getTestBooks();
     }
 
     @Override
     public Stream<? extends Arguments> provideArguments(ExtensionContext extensionContext){
-        var names = List.of("Book with 1 author and 1 genre",
-                            "Book with 2 authors and 1 genre",
-                            "Book with 2 authors and 2 genres");
+        var names = List.of("Book with no comments",
+                            "Book with 1 comment",
+                            "Book with 2 comments");
         return IntStream.range(0, this.books.size()).mapToObj(i -> Arguments.of(names.get(i), books.get(i)));
     }
 }
