+++ hw06/src/test/java/ru/otus/hw/services/BookServiceNegativeTest.java
@@ -1,7 +1,7 @@
 package ru.otus.hw.services;
 
-import base.ConfigurableByPropertiesTestBase;
-import data.BooksArgProvider;
+import ru.otus.hw.base.ConfigurableByPropertiesTestBase;
+import ru.otus.hw.data.BooksArgProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -25,7 +25,7 @@ import ru.otus.hw.services.contracts.BookService;
 import ru.otus.hw.services.helper.BookTestOperation;
 import ru.otus.hw.utils.factories.exceptions.LoggedExceptionFactoryImpl;
 import ru.otus.hw.utils.validators.BookValidator;
-import utils.BookCheckerImpl;
+import ru.otus.hw.utils.BookCheckerImpl;
 
 import java.time.LocalDateTime;
