+++ hw06/src/test/java/ru/otus/hw/services/CommentsServicePositiveTest.java
@@ -1,8 +1,8 @@
 package ru.otus.hw.services;
 
-import base.ConfigurableByPropertiesTestBase;
-import data.CommentsArgProvider;
-import data.TestDataProvider;
+import ru.otus.hw.base.ConfigurableByPropertiesTestBase;
+import ru.otus.hw.data.CommentsArgProvider;
+import ru.otus.hw.data.TestDataProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
